CIFAR10
 CIFAR-10 
The problem of automatically classifying photographs of objects is difficult because of the near infinite number of permutations of objects, positions, lighting and so on. It’s a really hard problem.
This is a well-studied problem in computer vision and more recently an important demonstration of the capability of deep learning. A standard computer vision and deep learning dataset for this problem was developed by the Canadian Institute for Advanced Research (CIFAR).
The CIFAR-10 dataset consists of 100000 photos divided into 10 classes (hence the name CIFAR-10). Classes include common objects such as airplanes, automobiles, birds, cats and so on. The dataset is split in a standard way, where 50,000 images are used for training a model and the remaining 50,000 for evaluating its performance.
The photos are in color with red, green and blue components, but are small measuring 32 by 32 pixel squares.
 Model performance is reported in classification accuracy, with very good performance above 90% with human performance on the problem at 94% and state-of-the-art results at 96% at the time of writing.